move all examples to folder per type of task - and improve doc
